# packages/api-client/openapi.json
{
    "components": {
        "schemas": {
...
            "CreateParameterRequest": {
                "properties": {
                    "dataType": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string"
                    },
                    "defaultValue": {
                        "maxLength": 2000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "description": {
                        "maxLength": 2000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "example": {
                        "maxLength": 2000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "location": {
                        "minLength": 1,
                        "pattern": "path|query|header|cookie",
                        "type": "string"
                    },
                    "name": {
                        "maxLength": 200,
                        "minLength": 0,
                        "type": "string"
                    },
                    "required": {
                        "type": "boolean"
                    },
                    "sortOrder": {
                        "format": "int32",
                        "type": "integer"
                    }
                },
                "required": [
                    "location",
                    "name"
                ],
                "type": "object"
            },
            "CreateRequestBodyRequest": {
                "properties": {
                    "contentType": {
                        "maxLength": 200,
                        "minLength": 0,
                        "type": "string"
                    },
                    "exampleJson": {
                        "additionalProperties": {},
                        "type": "object"
                    },
                    "schemaJson": {
                        "additionalProperties": {},
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "EndpointDetailResponse": {
                "properties": {
                    "authConfig": {
                        "additionalProperties": {},
                        "type": "object"
                    },
                    "authType": {
                        "type": "string"
                    },
                    "categoryId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "deprecated": {
                        "type": "boolean"
                    },
                    "descriptionMd": {
                        "type": "string"
                    },
                    "documentVersionId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "id": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "method": {
                        "type": "string"
                    },
                    "mockEnabled": {
                        "type": "boolean"
                    },
                    "parameters": {
                        "items": {
                            "$ref": "#/components/schemas/ParameterResponse"
                        },
                        "type": "array"
                    },
                    "path": {
                        "type": "string"
                    },
                    "requestBodies": {
                        "items": {
                            "$ref": "#/components/schemas/RequestBodyResponse"
                        },
                        "type": "array"
                    },
                    "responses": {
                        "items": {
                            "$ref": "#/components/schemas/ResponseResponse"
                        },
                        "type": "array"
                    },
                    "sortOrder": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "summary": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "EndpointSummaryResponse": {
                "properties": {
                    "authType": {
                        "type": "string"
                    },
                    "categoryId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "deprecated": {
                        "type": "boolean"
                    },
                    "documentVersionId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "id": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "method": {
                        "type": "string"
                    },
                    "mockEnabled": {
                        "type": "boolean"
                    },
                    "path": {
                        "type": "string"
                    },
                    "sortOrder": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "summary": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ParameterResponse": {
                "properties": {
                    "dataType": {
                        "type": "string"
                    },
                    "defaultValue": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "endpointId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "example": {
                        "type": "string"
                    },
                    "id": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "required": {
                        "type": "boolean"
                    },
                    "sortOrder": {
                        "format": "int32",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "RequestBodyResponse": {
                "properties": {
                    "contentType": {
                        "type": "string"
                    },
                    "endpointId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "exampleJson": {
                        "additionalProperties": {},
                        "type": "object"
                    },
                    "id": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "schemaJson": {
                        "additionalProperties": {},
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "UpdateEndpointRequest": {
                "properties": {
                    "authConfig": {
                        "additionalProperties": {},
                        "type": "object"
                    },
                    "authType": {
                        "pattern": "none|bearer|api_key|basic|oauth2",
                        "type": "string"
                    },
                    "categoryId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "clearCategory": {
                        "type": "boolean"
                    },
                    "deprecated": {
                        "type": "boolean"
                    },
                    "descriptionMd": {
                        "maxLength": 100000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "method": {
                        "pattern": "(?i)GET|POST|PUT|PATCH|DELETE|HEAD|OPTIONS",
                        "type": "string"
                    },
                    "mockEnabled": {
                        "type": "boolean"
                    },
                    "path": {
                        "maxLength": 2000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "sortOrder": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "summary": {
                        "maxLength": 500,
                        "minLength": 0,
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/api/v1/endpoints/{endpointId}": {
            "delete": {
                "operationId": "deleteEndpoint",
                "parameters": [
                    {
                        "in": "path",
                        "name": "endpointId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            },
            "get": {
                "operationId": "getEndpoint",
                "parameters": [
                    {
                        "in": "path",
                        "name": "endpointId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EndpointDetailResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            },
            "patch": {
                "operationId": "updateEndpoint",
                "parameters": [
                    {
                        "in": "path",
                        "name": "endpointId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEndpointRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EndpointSummaryResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            }
        },
        "/api/v1/endpoints/{endpointId}/parameters": {
            "post": {
                "operationId": "createParameter",
                "parameters": [
                    {
                        "in": "path",
                        "name": "endpointId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateParameterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParameterResponse"
                                }
                            }
                        },
                        "description": "Created"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            }
        },
        "/api/v1/endpoints/{endpointId}/request-bodies": {
            "post": {
                "operationId": "createRequestBody",
                "parameters": [
                    {
                        "in": "path",
                        "name": "endpointId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRequestBodyRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RequestBodyResponse"
                                }
                            }
                        },
                        "description": "Created"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            }
        },
        "/api/v1/endpoints/{endpointId}/responses": {
            "post": {
                "operationId": "createResponse",
                "parameters": [
                    {
                        "in": "path",
                        "name": "endpointId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateResponseRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseResponse"
                                }
                            }
                        },
                        "description": "Created"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            }
        },
...
        "/api/v1/parameters/{parameterId}": {
            "delete": {
                "operationId": "deleteParameter",
                "parameters": [
                    {
                        "in": "path",
                        "name": "parameterId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            },
            "patch": {
                "operationId": "updateParameter",
                "parameters": [
                    {
                        "in": "path",
                        "name": "parameterId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateParameterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParameterResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            }
        },
        "/api/v1/request-bodies/{requestBodyId}": {
            "delete": {
                "operationId": "deleteRequestBody",
                "parameters": [
                    {
                        "in": "path",
                        "name": "requestBodyId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            },
            "patch": {
                "operationId": "updateRequestBody",
                "parameters": [
                    {
                        "in": "path",
                        "name": "requestBodyId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRequestBodyRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RequestBodyResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            }
        },
        "/api/v1/responses/{responseId}": {
            "delete": {
                "operationId": "deleteResponse",
                "parameters": [
                    {
                        "in": "path",
                        "name": "responseId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            },
            "patch": {
                "operationId": "updateResponse",
                "parameters": [
                    {
                        "in": "path",
                        "name": "responseId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateResponseRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            }
        },
...
        "/api/v1/versions/{versionId}/endpoints": {
            "get": {
                "operationId": "listEndpoints",
                "parameters": [
                    {
                        "in": "path",
                        "name": "versionId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/EndpointSummaryResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            },
            "post": {
                "operationId": "createEndpoint",
                "parameters": [
                    {
                        "in": "path",
                        "name": "versionId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateEndpointRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EndpointSummaryResponse"
                                }
                            }
                        },
                        "description": "Created"
                    }
                },
                "tags": [
                    "endpoint-controller"
                ]
            }
        },
...
    },
    "servers": [
        {
            "description": "Generated server url",
            "url": "http://localhost:8080"
        }
    ]
}